moved vis specific ui to vis packages
