+++ src/vis/bar/BarChart.tsx
