+++ src/vis/bar/BarVis.tsx
