+++ src/vis/bar/BarVisSidebar.tsx
@@ -4,9 +4,9 @@ import * as React from 'react';
 import { useMemo } from 'react';
 import { ColumnInfo, EAggregateTypes, EColumnTypes, ICommonVisSideBarProps } from '../interfaces';
 import { AggregateTypeSelect } from '../sidebar/AggregateTypeSelect';
-import { BarDirectionButtons } from '../sidebar/BarDirectionButtons';
+import { BarDirectionButtons } from './BarDirectionButtons';
 import { FilterButtons } from '../sidebar/FilterButtons';
-import { GroupSelect } from '../sidebar/GroupSelect';
+import { GroupSelect } from './GroupSelect';
 import { SingleColumnSelect } from '../sidebar/SingleColumnSelect';
 import { VisTypeSelect } from '../sidebar/VisTypeSelect';
 import { EBarDirection, EBarDisplayType, EBarGroupingType, IBarConfig } from './interfaces';
