+++ src/vis/bar/SingleBarChart.tsx
