+++ src/vis/bar/barComponents/Legend.tsx
