+++ src/vis/bar/barComponents/SingleBar.tsx
