+++ src/vis/bar/barTypes/SimpleBars.tsx
