+++ src/vis/bar/hooks/useGetBarScales.ts
