+++ src/vis/bar/interfaces.ts
