+++ src/vis/hexbin/HexSizeSlider.tsx
