+++ src/vis/hexbin/HexSizeSwitch.tsx
