+++ src/vis/raincloud/AggregateRainSwitch.tsx
