+++ src/vis/scatter/OpacitySlider.tsx
