+++ src/vis/scatter/ScatterVisSidebar.tsx
@@ -3,10 +3,10 @@ import merge from 'lodash/merge';
 import * as React from 'react';
 import { useMemo } from 'react';
 import { ColumnInfo, EColumnTypes, ENumericalColorScaleType, ESupportedPlotlyVis, ICommonVisSideBarProps } from '../interfaces';
-import { ColorSelect } from '../sidebar/ColorSelect';
+import { ColorSelect } from './ColorSelect';
 import { FilterButtons } from '../sidebar/FilterButtons';
 import { NumericalColumnSelect } from '../sidebar/NumericalColumnSelect';
-import { OpacitySlider } from '../sidebar/OpacitySlider';
+import { OpacitySlider } from './OpacitySlider';
 import { SingleColumnSelect } from '../sidebar/SingleColumnSelect';
 import { VisTypeSelect } from '../sidebar/VisTypeSelect';
 import { IScatterConfig } from './interfaces';
