+++ src/vis/sidebar/index.ts
@@ -1,14 +1,14 @@
-export * from './BarDirectionButtons';
-export * from './BarDisplayTypeButtons';
-export * from './BarGroupTypeButtons';
+export * from '../bar/BarDirectionButtons';
+export * from '../bar/BarDisplayTypeButtons';
+export * from '../bar/BarGroupTypeButtons';
 export * from './BrushOptionButtons';
 export * from './CategoricalColumnSelect';
-export * from './ColorSelect';
+export * from '../scatter/ColorSelect';
 export * from './FilterButtons';
-export * from './GroupSelect';
+export * from '../bar/GroupSelect';
 export * from './NumericalColorButtons';
 export * from './NumericalColumnSelect';
-export * from './OpacitySlider';
+export * from '../scatter/OpacitySlider';
 export * from './utils';
-export * from './ViolinOverlayButtons';
+export * from '../violin/ViolinOverlayButtons';
 export * from './VisTypeSelect';
