+++ src/vis/violin/ViolinOverlayButtons.tsx
@@ -1,6 +1,6 @@
 import { Container, SegmentedControl, Stack, Text } from '@mantine/core';
 import * as React from 'react';
-import { EViolinOverlay } from '../violin/interfaces';
+import { EViolinOverlay } from './interfaces';
 
 interface ViolinOverlayProps {
   callback: (s: EViolinOverlay) => void;
